@@ -1,5 +1,5 @@
 {
-  "name": "toshokan",
+  "name": "toshokan-new",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,27 +8,37 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@headlessui/vue": "^1.2.0",
+    "@heroicons/vue": "^1.0.1",
+    "@tailwindcss/aspect-ratio": "^0.2.1",
+    "@tailwindcss/forms": "^0.3.3",
+    "@vuelidate/core": "^2.0.0-alpha.19",
+    "@vuelidate/validators": "^2.0.0-alpha.17",
+    "apexcharts": "^3.27.1",
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
+    "fuse.js": "^6.4.6",
     "lodash.groupby": "^4.6.0",
+    "lodash.orderby": "^4.6.0",
     "lodash.uniqby": "^4.7.0",
-    "nanoid": "^3.1.20",
+    "nanoid": "^3.1.23",
+    "paginator": "^1.0.0",
     "register-service-worker": "^1.7.1",
-    "slugify": "^1.4.6",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vuelidate": "^0.7.6",
-    "vuetify": "^2.4.0",
-    "vuex": "^3.4.0"
+    "slugify": "^1.5.3",
+    "tailwindcss-interaction-variants": "^5.0.0",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0",
+    "vue3-apexcharts": "^1.4.0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.3.3",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-pwa": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-plugin-vuex": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-standard": "^5.1.2",
     "autoprefixer": "^9",
     "babel-eslint": "^10.1.0",
@@ -37,15 +47,9 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.12.0",
+    "eslint-plugin-vue": "^7.0.0",
+    "git-describe": "^4.0.4",
     "postcss": "^7",
-    "sass": "^1.32.0",
-    "sass-loader": "^10.0.0",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "tailwindcss-interaction-variants": "^5.0.0",
-    "vue-cli-plugin-vuetify": "~2.1.0",
-    "vue-template-compiler": "^2.6.11",
-    "vuetify-loader": "^1.7.0"
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
   }
 }
