@@ -1,19 +1,12 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import './registerServiceWorker'
 import router from './router'
 import store from './store'
-import vuetify from './plugins/vuetify'
-import Vuelidate from 'vuelidate'
+
 import './index.css'
 
-Vue.config.productionTip = false
-
-Vue.use(Vuelidate)
-
-new Vue({
-  router,
-  store,
-  vuetify,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App)
+  .use(store)
+  .use(router)
+  .mount('#app')
