@@ -1,9 +1,5 @@
 <template>
-  <section class="pa-md-2">
-    <v-fade-transition mode="out-in">
-      <router-view/>
-    </v-fade-transition>
-  </section>
+  <router-view/>
 </template>
 
 <script>
