@@ -1,162 +1,51 @@
 <template>
-  <div class="ma-3">
-    <v-row v-if="!loading">
-      <v-col md="6" cols="12">
-        <v-card>
-          <v-card-title class="text-body-1">
-            Gasto mensal
-          </v-card-title>
+  <div>
+    <div class="h-full flex flex-col">
+      <SimpleHeader title="Estatísticas" :loading="loading" />
 
-          <v-sheet class="pa-2">
-            <g-line-chart
-              :columns="monthlyColumns"
-              :rows="monthlyRows"
-              ref="monthlyChart"
-            />
-          </v-sheet>
-        </v-card>
-      </v-col>
+      <div class="max-w-7xl w-full mx-auto py-6 px-6 lg:px-8 space-y-6">
+        <div class="space-y-6 md:space-y-0 md:grid md:grid-cols-2 gap-6">
+          <MonthlyExpenseChart />
+          <MonthlyBoughtsChart />
+        </div>
+      </div>
 
-      <v-col md="6" cols="12">
-        <v-card>
-          <v-card-text>
-            <div class="text-body-1 font-weight-medium">
-              Itens comprados
-            </div>
-          </v-card-text>
-
-          <v-sheet class="pa-2">
-            <g-line-chart
-              :columns="monthlyColumns"
-              :rows="countRows"
-              :chart-area="{ left: 40, width: '89%' }"
-              :v-axis="{ format: 'decimal' }"
-              ref="countChart"
-            />
-          </v-sheet>
-        </v-card>
-      </v-col>
-    </v-row>
-
-    <v-row v-if="!loading">
-      <v-col
-        v-for="mostPresent in mostPresents"
-        :key="mostPresent.key"
-        md="4"
-        cols="12"
-      >
-        <v-card>
-          <v-card-title class="text-body-1">
-            {{ mostPresent.title }}
-          </v-card-title>
-          <v-simple-table>
-            <thead>
-              <tr>
-                <th class="text-left">
-                  {{ mostPresent.value }}
-                </th>
-                <th class="text-right">
-                  {{ mostPresent.count }}
-                </th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr
-                v-for="row in stats[mostPresent.key]"
-                :key="row.name"
-              >
-                <td>
-                  <router-link
-                    :to="{ name: 'collection', query: { search: row.name } }"
-                  >
-                    {{ row.name }}
-                  </router-link>
-                </td>
-                <td class="text-right">{{ row.count }}</td>
-              </tr>
-            </tbody>
-          </v-simple-table>
-        </v-card>
-      </v-col>
-    </v-row>
+      <div v-if="!loading" class="max-w-7xl w-full mx-auto pb-6 px-6 lg:px-8 flex-1">
+        <div class="h-96 md:h-full flex flex-col items-center justify-center border-4 border-dashed border-gray-200 dark:border-gray-700 rounded-lg">
+          <DotsHorizontalIcon class="w-10 h-10 text-gray-400 dark:text-gray-600" aria-hidden="true" />
+          <p class="text-lg text-gray-400 dark:text-gray-600">Mais estatísticas em breve.</p>
+        </div>
+      </div>
+    </div>
   </div>
 </template>
 
 <script>
-import { mapMutations, mapState } from 'vuex'
+import { computed } from 'vue'
+import { useStore } from 'vuex'
 
-import GLineChart from '@/components/GLineChart'
+import { DotsHorizontalIcon } from '@heroicons/vue/solid'
+
+import MonthlyBoughtsChart from '@/components/MonthlyBoughtsChart'
+import MonthlyExpenseChart from '@/components/MonthlyExpenseChart'
+import SimpleHeader from '@/components/SimpleHeader'
 
 export default {
   name: 'DashboardStats',
 
   components: {
-    GLineChart
+    MonthlyBoughtsChart,
+    MonthlyExpenseChart,
+    SimpleHeader,
+    DotsHorizontalIcon
   },
 
-  data: () => ({
-    mostPresents: [
-      {
-        title: 'Séries mais presentes',
-        value: 'Série',
-        count: 'Volumes',
-        key: 'series'
-      },
-      {
-        title: 'Autores mais presentes',
-        value: 'Autor',
-        count: 'Volumes',
-        key: 'authors'
-      },
-      {
-        title: 'Editoras mais presentes',
-        value: 'Editora',
-        count: 'Volumes',
-        key: 'imprints'
-      }
-    ]
-  }),
+  setup () {
+    const store = useStore()
 
-  computed: {
-    monthlyColumns () {
-      return [
-        { type: 'string', title: 'Período' },
-        { type: 'number', title: 'Valor' }
-      ]
-    },
+    const loading = computed(() => store.state.sheet.loading)
 
-    monthlyRows () {
-      return this.stats.monthly.map(m => [
-        m.month,
-        parseFloat(
-          m.totalSpent
-            .replace('R$ ', '')
-            .replace(/\./g, '')
-            .replace(',', '.')
-        )
-      ])
-    },
-
-    countRows () {
-      return this.stats.monthly.map(m => [
-        m.month,
-        parseFloat(
-          m.count
-            .replace(/\./g, '')
-            .replace(',', '.')
-        )
-      ])
-    },
-
-    ...mapState('sheet', ['loading', 'stats'])
-  },
-
-  methods: {
-    ...mapMutations('appbar', ['updateIcons'])
-  },
-
-  mounted: function () {
-    this.updateIcons([])
+    return { loading }
   }
 }
 </script>
